--- RevitCSV/RevitCSV.R2025/RevitCSV.2025.addin
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<RevitAddIns>
-  <AddIn Type="Application">
-    <Name>RevitCSV</Name>
-    <FullClassName>RevitCSV.ExternalApplication</FullClassName>
-    <Text>RevitCSV</Text>
-    <Description>RevitCSV</Description>
-    <VisibilityMode>AlwaysVisible</VisibilityMode>
-    <Assembly>RevitCSV.R2025.dll</Assembly>
-    <AddInId>
-      72158e33-90b1-4a9e-af03-1d821326c2fe
-    </AddInId>
-    <VendorId>ADSK</VendorId>
-    <VendorDescription>Autodesk, Inc, www.autodesk.com</VendorDescription>
-  </AddIn>
-</RevitAddIns>
