+++ RevitCSV/RevitCSV.Services/Services/Collector/IElementCollector.cs
@@ -1,0 +1,10 @@
+﻿using Autodesk.Revit.DB;
+using System.Collections.Generic;
+
+namespace Services
+{
+    public interface IElementCollector
+    {
+        List<Element> FetchElements();
+    }
+}
