+++ RevitCSV/RevitCSV.Services/Services/Exporter/IExporter.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Services.Exporter
+{
+    public interface IExporter
+    {   
+        void Export(TabularData data, string path);
+    }
+}
