+++ RevitCSV/RevitCSV.sln
@@ -3,14 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV", "RevitCSV\RevitCSV.csproj", "{72158E33-90B1-4A9E-AF03-1D821326C2FE}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.Installer", "RevitCSV.Installer\RevitCSV.Installer.csproj", "{590DF1CE-98EB-4FAB-BAA2-AF3313ABD621}"
 	ProjectSection(ProjectDependencies) = postProject
 		{50EC8192-F4AB-4388-BCBB-CE6CBDD330F0} = {50EC8192-F4AB-4388-BCBB-CE6CBDD330F0}
-		{72158E33-90B1-4A9E-AF03-1D821326C2FE} = {72158E33-90B1-4A9E-AF03-1D821326C2FE}
-		{9601561B-6C20-47A8-BA5D-76939E47166C} = {9601561B-6C20-47A8-BA5D-76939E47166C}
 		{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386} = {9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}
+		{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90} = {CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}
 		{D42C0456-29DD-4101-96C1-9BD64BBD0C20} = {D42C0456-29DD-4101-96C1-9BD64BBD0C20}
 		{EAA8E6CC-D76A-4583-A73C-6B9221C228F4} = {EAA8E6CC-D76A-4583-A73C-6B9221C228F4}
 	EndProjectSection
@@ -21,9 +18,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.R2023", "RevitCSV.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.R2024", "RevitCSV.R2024\RevitCSV.R2024.csproj", "{EAA8E6CC-D76A-4583-A73C-6B9221C228F4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.R2025", "RevitCSV.R2025\RevitCSV.R2025.csproj", "{9601561B-6C20-47A8-BA5D-76939E47166C}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.R2021", "RevitCSV.R2021\RevitCSV.R2021.csproj", "{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevitCSV.Common", "RevitCSV.Common\RevitCSV.Common.csproj", "{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "RevitCSV.Services", "RevitCSV.Services\RevitCSV.Services.shproj", "{7BAD5848-8DEA-4FD1-A3B1-738414AB4303}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,10 +30,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{72158E33-90B1-4A9E-AF03-1D821326C2FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72158E33-90B1-4A9E-AF03-1D821326C2FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72158E33-90B1-4A9E-AF03-1D821326C2FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72158E33-90B1-4A9E-AF03-1D821326C2FE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{590DF1CE-98EB-4FAB-BAA2-AF3313ABD621}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{590DF1CE-98EB-4FAB-BAA2-AF3313ABD621}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{590DF1CE-98EB-4FAB-BAA2-AF3313ABD621}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -51,19 +46,26 @@ Global
 		{EAA8E6CC-D76A-4583-A73C-6B9221C228F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EAA8E6CC-D76A-4583-A73C-6B9221C228F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EAA8E6CC-D76A-4583-A73C-6B9221C228F4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9601561B-6C20-47A8-BA5D-76939E47166C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9601561B-6C20-47A8-BA5D-76939E47166C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9601561B-6C20-47A8-BA5D-76939E47166C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9601561B-6C20-47A8-BA5D-76939E47166C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9A1A1DDB-B1F6-4EC5-937C-D0163A8F5386}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB40EFB9-54E4-4DD1-8BC3-567BE5794F90}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5EE22718-020A-45FD-8DF2-73F005DDB58A}
+	EndGlobalSection
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		RevitCSV.Services\RevitCSV.Services.projitems*{50ec8192-f4ab-4388-bcbb-ce6cbdd330f0}*SharedItemsImports = 4
+		RevitCSV.Services\RevitCSV.Services.projitems*{7bad5848-8dea-4fd1-a3b1-738414ab4303}*SharedItemsImports = 13
+		RevitCSV.Services\RevitCSV.Services.projitems*{9a1a1ddb-b1f6-4ec5-937c-d0163a8f5386}*SharedItemsImports = 4
+		RevitCSV.Services\RevitCSV.Services.projitems*{d42c0456-29dd-4101-96c1-9bd64bbd0c20}*SharedItemsImports = 4
+		RevitCSV.Services\RevitCSV.Services.projitems*{eaa8e6cc-d76a-4583-a73c-6b9221c228f4}*SharedItemsImports = 4
 	EndGlobalSection
 EndGlobal
