--- RevitCSV/RevitCSV/Ribbon/PackageContents.xml
@@ -1,27 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ApplicationPackage SchemaVersion="1.0"
-                    AutodeskProduct="Revit"
-                    ProductType="Application"
-                    Name="Revit CSV"
-                    AppVersion="1.0.0.0"
-                    Description="CSV Export"
-                    Author="Libo, Fran"
-                    Icon=""
-                    AppNameSpace=""
-                    OnlineDocumentation=""
-                    HelpFile=""
-                    ProductCode=""
-                    UpgradeCode=""
-                    SupportedLocales="Enu"
-                    FriendlyVersion="1.0.0.0">
-	<Components Description="2023">
-		<RuntimeRequirements OS="Win64"
-							 Platform="Revit"
-							 SeriesMin="R2023"
-							 SeriesMax="R2023" />
-		<ComponentEntry AppName="Revit CSV"
-						Version="2023"
-						ModuleName="./Contents/2023/RevitCSV.addin"
-						AppDescription="RevitCSV" />
-	</Components>
-</ApplicationPackage>
