--- RevitCSV/RevitCSV/Services/Collector/IElementCollector.cs
@@ -1,10 +1,0 @@
-﻿using Autodesk.Revit.DB;
-using System.Collections.Generic;
-
-namespace Services
-{
-    public interface IElementCollector
-    {
-        List<Element> FetchElements();
-    }
-}
