--- RevitCSV/RevitCSV/Services/Exporter/IExporter.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Services.Exporter
-{
-    public interface IExporter
-    {   
-        void Export(TabularData data, string path);
-    }
-}
