--- RevitCSV/RevitCSV/Services/Exporter/TabularData.cs
@@ -1,7 +1,0 @@
-﻿using System.Collections.Generic;
-
-public class TabularData
-{
-    public List<string> Headers { get; set; } = new List<string>();
-    public List<List<string>> Rows { get; set; } = new List<List<string>>();
-}
