--- RevitCSV/RevitCSV/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Revit.RevitApi.x64" version="2023.0.0" targetFramework="net48" />
-  <package id="Revit.RevitApiUI.x64" version="2023.0.0" targetFramework="net48" />
-  <package id="Revit_All_Main_Versions_API_x64" version="2023.0.0" targetFramework="net48" />
-</packages>
